requirements.txt
django-cms>=3.8,<3.9
djangocms-admin-style>=2.0,<3.0
django-treebeard>=4.0,<5.0

djangocms-text-ckeditor>=4.0,<5.0
djangocms-link>=3.0,<4.0
djangocms-icon>=2.0,<3.0
djangocms-style>=3.0,<4.0
djangocms-googlemap>=2.0,<3.0
djangocms-video>=3.0,<4.0
djangocms-file>=3.0,<4.0
djangocms-picture>=3.0,<4.0
djangocms-bootstrap4>=2.0,<3.0
easy_thumbnails
django-filer>=1.3
Django<3.2
django-classy-tags>=2.0
django-sekizai>=2.0
django-mptt>0.9
html5lib>=1.0.1
Pillow>=3.0
six
pytz

django-crispy-forms==1.10.0
django-tinymce==3.2.0
django-bleach==0.6.1
django-cleanup==5.1.0
django-sendfile2==0.6.0
django-debug-toolbar==3.2
django-select2==7.6.1
social-auth-app-django==4.0.0
django-admin-sortable2==0.7.8
django-imagekit==4.0.2
python-dateutil==2.8.1
Faker==5.8.0
psycopg2==2.8.6

mock==4.0.3
freezegun==1.1.0
coverage==5.4
django-coverage-plugin==1.8.0

mypy==0.800
django-stubs==1.7.0